backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20

sqlalchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10

redis==6.4.0
celery==5.5.3

httpx==0.28.1
beautifulsoup4==4.13.4
tenacity==9.1.2

scikit-learn==1.7.1
spacy==3.8.7
PyMuPDF==1.26.4
python-docx==1.2.0

python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

python-json-logger==3.3.0

pytest==8.4.1
pytest-cov==6.2.1
pytest-asyncio==1.1.0
ruff==0.12.10
mypy==1.17.1
